integration-tests/pytest/pyproject.toml
[project]
name = "draft-integration-tests"
version = "0.1.0"
description = "Integration tests for the Draft projects"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "httpx>=0.28.0",
    "mypy>=1.20.0",
    "openapi-core>=0.23.1",
    "pytest-asyncio>=1.3.0",
    "pytest-datadir>=1.8.0",
    "pytest-httpserver>=1.1.5",
    "pytest>=9.0.3",
    "pyyaml>=6.0",
    "ruff>=0.15.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]

[tool.ruff.lint.isort]
known-first-party = ["draft_*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["tests"]

[tool.mypy]
check_untyped_defs = true

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20260408",
]
